fixed checkstyle issues in adapters for search by flight number and search by arrival airport
