+++ src/java/adapters/search_by_arrival_airport/SearchByArrivalAirportController.java
@@ -1,10 +1,10 @@
-package adapters.SearchByArrivalAirport;
+package adapters.search_by_arrival_airport;
 
 import use_case.SearchByArrivalAirport.SearchByArrivalAirportInputBoundary;
 import use_case.SearchByArrivalAirport.SearchByArrivalAirportInputData;
 
 /**
- * The controller for the Search By Airport Use Case.
+ * The controller for the Search By Arrival Airport Use Case.
  */
 public class SearchByArrivalAirportController {
 
@@ -15,15 +15,17 @@ public class SearchByArrivalAirportController {
     }
 
     /**
-     * Handles the user request to search for flights by airport
-     * @param airportCode the airport IATA code entered by the user
+     * Handles the user request to search for flights by arrival airport.
+     * @param airportCode the arrival airport IATA code entered by the user
+     * @throws IllegalArgumentException when the Airport IATA code is empty
      */
     public void execute(String airportCode) {
+        // Checks if the airportCode is empty and if it is, throws a IllegalArgumentException
         if (airportCode == null || airportCode.trim().isEmpty()) {
             throw new IllegalArgumentException("Airport IATA code cannot be empty.");
         }
 
-        // Create input data and call the input boundary
+        // Creates input data and then excecutes the Interactor with the input boundary
         SearchByArrivalAirportInputData inputData = new SearchByArrivalAirportInputData(airportCode.trim());
         searchByArrivalAirportInteractor.execute(inputData);
     }
