+++ src/java/adapters/search_by_flight_number/SearchByFlightNumberController.java
@@ -1,9 +1,11 @@
-package adapters.SearchByFlightNumber;
+package adapters.search_by_flight_number;
 
 import use_case.SearchByFlightNumber.SearchByFlightNumberInputBoundary;
 import use_case.SearchByFlightNumber.SearchByFlightNumberInputData;
 
-
+/**
+ * The controller for the Search By Flight Number Use Case.
+ */
 public class SearchByFlightNumberController {
     private final SearchByFlightNumberInputBoundary searchByFlightNumberInteractor;
 
@@ -13,9 +15,16 @@ public class SearchByFlightNumberController {
 
     /**
      * Executes the search by flight number use case.
-     * @param flightNumber
+     * @param flightNumber the flight number entered by the user
+     * @throws IllegalArgumentException when the Airport IATA code is empty
      */
     public void execute(String flightNumber) {
+        // Checks if the flight number is empty and if it is, throws a IllegalArgumentException
+        if (flightNumber == null) {
+            throw new IllegalArgumentException("Flight number cannot be null");
+        }
+
+        // Creates input data and then excecutes the Interactor with the input boundary
         final SearchByFlightNumberInputData inputData = new SearchByFlightNumberInputData(flightNumber);
         this.searchByFlightNumberInteractor.execute(inputData);
     }
