+++ src/java/adapters/search_by_flight_number/SearchByFlightNumberState.java
@@ -1,9 +1,9 @@
-package adapters.SearchByFlightNumber;
+package adapters.search_by_flight_number;
 
 import adapters.AbstractState;
 
 /**
- * The state for the Search By Number View Model.
+ * The state for the Search By Flight Number View Model.
  */
 public class SearchByFlightNumberState extends AbstractState {
 }
